add String.fromCharacters
