+++ runtime/sema/type.go
@@ -3330,88 +3330,6 @@ func numberFunctionArgumentExpressionsChecker(targetType Type) ArgumentExpressio
 	}
 }
 
-func init() {
-
-	// Declare a function for the string type.
-	// For now, it has no parameters and creates an empty string
-
-	typeName := StringType.String()
-
-	// Check that the function is not accidentally redeclared
-
-	if BaseValueActivation.Find(typeName) != nil {
-		panic(errors.NewUnreachableError())
-	}
-
-	functionType := &FunctionType{
-		ReturnTypeAnnotation: NewTypeAnnotation(StringType),
-	}
-
-	addMember := func(member *Member) {
-		if functionType.Members == nil {
-			functionType.Members = &StringMemberOrderedMap{}
-		}
-		name := member.Identifier.Identifier
-		if functionType.Members.Contains(name) {
-			panic(errors.NewUnreachableError())
-		}
-		functionType.Members.Set(name, member)
-	}
-
-	addMember(NewUnmeteredPublicFunctionMember(
-		functionType,
-		StringTypeEncodeHexFunctionName,
-		StringTypeEncodeHexFunctionType,
-		StringTypeEncodeHexFunctionDocString,
-	))
-
-	addMember(NewUnmeteredPublicFunctionMember(
-		functionType,
-		StringTypeFromUtf8FunctionName,
-		StringTypeFromUtf8FunctionType,
-		StringTypeFromUtf8FunctionDocString,
-	))
-
-	BaseValueActivation.Set(
-		typeName,
-		baseFunctionVariable(
-			typeName,
-			functionType,
-			"Creates an empty string",
-		),
-	)
-}
-
-var StringTypeEncodeHexFunctionType = &FunctionType{
-	Parameters: []*Parameter{
-		{
-			Label:      ArgumentLabelNotRequired,
-			Identifier: "data",
-			TypeAnnotation: NewTypeAnnotation(
-				ByteArrayType,
-			),
-		},
-	},
-	ReturnTypeAnnotation: NewTypeAnnotation(
-		StringType,
-	),
-}
-
-var StringTypeFromUtf8FunctionType = &FunctionType{
-	Parameters: []*Parameter{
-		{
-			Label:          ArgumentLabelNotRequired,
-			Identifier:     "bytes",
-			TypeAnnotation: NewTypeAnnotation(ByteArrayType),
-		},
-	},
-	ReturnTypeAnnotation: NewTypeAnnotation(
-		&OptionalType{
-			Type: StringType,
-		},
-	),
-}
-
 func pathConversionFunctionType(pathType Type) *FunctionType {
 	return &FunctionType{
 		Parameters: []*Parameter{
